requirements.txt
tensorflow>=2.10
pandas>=1.5
numpy>=1.23
scikit-learn>=1.2
matplotlib>=3.7
fastapi>=0.100
uvicorn>=0.23
python-multipart>=0.0.6
pydantic>=2.0
python-dotenv>=1.0
pyyaml>=6.0
sqlalchemy>=2.0
jinja2>=3.1
dvc>=3.0
pillow>=9.0